backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
deepgram-sdk>=3.0.0
groq
resend
markdown
yt-dlp>=2024.03.10
reportlab
arabic-reshaper
python-bidi